refactor: Restructure CLI and config into infrastructure directory
Move CLI command handlers, services, and config files to
src/infrastructure Update related import paths and package.json entries
to reflect new structure
