+++ package.json
@@ -2,10 +2,10 @@
   "name": "git-friday",
   "version": "0.11.1",
   "description": "An intelligent CLI tool that uses AI to analyze git commits and generate a clear, human-readable report.",
-  "main": "dist/src/cli/index.js",
+  "main": "dist/src/infrastructure/cli/index.js",
   "type": "module",
   "bin": {
-    "friday": "dist/src/cli/index.js"
+    "friday": "dist/src/infrastructure/cli/index.js"
   },
   "files": [
     "dist",
@@ -13,7 +13,7 @@
     "README.md"
   ],
   "scripts": {
-    "dev": "tsx --watch --env-file .env src/cli/index.ts",
+    "dev": "tsx --watch --env-file .env src/infrastructure/cli/index.ts",
     "start": "node dist/index.js",
     "build": "rimraf dist && tsc -p tsconfig.build.json && tsc-alias",
     "lint": "eslint .",
