+++ src/infrastructure/cli/index.ts
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
 import 'dotenv/config';
-import pkg from '../../package.json' with { type: 'json' };
+import pkg from '../../../package.json' with { type: 'json' };
 
 import { Command } from 'commander';
 
-import { createConfig } from '@/config.js';
+import { createConfig } from '@/infrastructure/config/config.js';
 import { createDiContainer } from '@/infrastructure/di/container.js';
-import { report } from '@/cli/commands/report.command.js';
+import { report } from '@/infrastructure/cli/commands/report.command.js';
 
 function main(): void {
   const appConfig = createConfig();
